PASSWORD RESET FUNCTIONALITY

+++ create_db.py
@@ -17,19 +17,22 @@ import os
 fake = Faker()
 
 if config.DB_TYPE == 'postgres':
-	conn = psycopg2.connect(host=config.PG_HOST, database=config.DATABASE, user=config.PG_USER, password=config.PG_PASSWORD)
-	conn.set_session(autocommit=True)
-	cur = conn.cursor()
-	logging.info('connected to {0} on database {1}'.format(config.PG_HOST, config.DATABASE))
-	cur.execute("DROP SCHEMA public CASCADE;")
-	cur.execute("CREATE SCHEMA public;")
-	# commented queries below throw errors about not existing
-	#cur.execute("GRANT ALL ON SCHEMA public TO postgres;")
-	cur.execute("GRANT ALL ON SCHEMA public TO public;")
-	cur.execute("COMMENT ON SCHEMA public IS 'standard public schema';")
-	#cur.execute("CREATE USER test WITH PASSWORD 'test';")
-	#cur.execute("ALTER SCHEMA public OWNER to postgres;")
-	logging.info('Succesfully provisioned database')
+	if config.USE_RECREATE:
+		conn = psycopg2.connect(host=config.PG_HOST, database=config.DATABASE, user=config.PG_USER, password=config.PG_PASSWORD)
+		conn.set_session(autocommit=True)
+		cur = conn.cursor()
+		logging.info('connected to {0} on database {1}'.format(config.PG_HOST, config.DATABASE))
+		cur.execute("DROP SCHEMA public CASCADE;")
+		cur.execute("CREATE SCHEMA public;")
+		# commented queries below throw errors about not existing
+		#cur.execute("GRANT ALL ON SCHEMA public TO postgres;")
+		cur.execute("GRANT ALL ON SCHEMA public TO public;")
+		cur.execute("COMMENT ON SCHEMA public IS 'standard public schema';")
+		#cur.execute("CREATE USER test WITH PASSWORD 'test';")
+		#cur.execute("ALTER SCHEMA public OWNER to postgres;")
+		logging.info('Succesfully provisioned database')
+	else:
+		os.system('bash recreate_psql_db.sh')
 elif config.DB_TYPE == 'sqlite':
 	try:
 		os.remove('{0}.db'.format(config.PG_USER))
